backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-multipart==0.0.17
pydantic==2.10.3
pydantic-settings==2.6.1
numpy==2.1.3
scikit-learn==1.5.2
google-generativeai>=0.8.0
python-dotenv==1.0.1
